requirements.txt
cryptography~=43.0.3

requests~=2.25.1
Flask~=3.1.0
pillow~=11.0.0
python-dotenv-1.1.0~=1.1.0